exercises/04_agent_routing/requirements.txt
temporalio
openai-agents
httpx
rich
pytz
